backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
python-email-validator==2.1.0
google-generativeai==0.8.3
asyncpg==0.29.0
